# apps/dashboard/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_api_v1_replay_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_replay_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplayRequest": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "anyOf": [
              {
                "maximum": 100000000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "path": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "speed": {
            "default": 0.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Speed",
            "type": "number"
          }
        },
        "required": [
          "path"
        ],
        "title": "ReplayRequest",
        "type": "object"
      },
...
      "ScenarioRequest": {
        "additionalProperties": false,
        "properties": {
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ]
            },
            "maxProperties": 10,
            "title": "Params",
            "type": "object"
          }
        },
        "title": "ScenarioRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nSentinelX network intrusion detection and prevention API.\n\n**Authentication.** Obtain tokens from `POST /api/v1/auth/login` and send\n`Authorization: Bearer <access_token>`. Browser sessions use httpOnly cookies and must\necho the `sx_csrf` cookie in an `X-CSRF-Token` header on state-changing requests.\n\n**Roles.** `viewer` reads; `analyst` triages, tests rules and runs replays; `admin`\nchanges rules, firewall state and settings.\n\n**Safety.** Prevention is off by default. Responses honour `DRY_RUN`, and every\nblock passes the safety guard whether it came from a detector or a person.\n",
    "title": "SentinelX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/replay": {
      "get": {
        "operationId": "recent_api_v1_replay_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Recent Api V1 Replay Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent",
        "tags": [
          "replay"
        ]
      },
      "post": {
        "operationId": "start_api_v1_replay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Api V1 Replay Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start",
        "tags": [
          "replay"
        ]
      }
    },
...
    "/api/v1/replay/files/inspect": {
      "get": {
        "operationId": "inspect_api_v1_replay_files_inspect_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Inspect Api V1 Replay Files Inspect Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect",
        "tags": [
          "replay"
        ]
      }
    },
...
    "/api/v1/replay/scenarios/{name}": {
      "post": {
        "description": "Generates a capture *file* for testing. Nothing is transmitted on the network.",
        "operationId": "generate_api_v1_replay_scenarios__name__post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Api V1 Replay Scenarios  Name  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write a synthetic scenario to a PCAP file",
        "tags": [
          "replay"
        ]
      }
    },
    "/api/v1/replay/upload": {
      "post": {
        "operationId": "upload_api_v1_replay_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_replay_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Api V1 Replay Upload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload",
        "tags": [
          "replay"
        ]
      }
    },
    "/api/v1/replay/{replay_id}": {
      "get": {
        "operationId": "get_api_v1_replay__replay_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "replay_id",
            "required": true,
            "schema": {
              "title": "Replay Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Api V1 Replay  Replay Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get",
        "tags": [
          "replay"
        ]
      }
    },
    "/api/v1/replay/{replay_id}/cancel": {
      "post": {
        "operationId": "cancel_api_v1_replay__replay_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "replay_id",
            "required": true,
            "schema": {
              "title": "Replay Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Api V1 Replay  Replay Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel",
        "tags": [
          "replay"
        ]
      }
    },
...
  }
}